requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.29
django-filter==24.3
django-parler==2.3
djangorestframework==3.15.2
pillow==10.4.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
sqlparse==0.5.5
typing_extensions==4.13.2
djangorestframework-simplejwt==5.3.1
pytz==2026.1
stripe==11.6.0
